refactor: relocate lib/quality/repair.py to lib/repair.py (#512)
The orphan/inconsistency-detection + repair-suggestion cluster
(OrphanInfo, RepairAction, SlskdOrphanTransfer, find_orphaned_downloads,
find_inconsistencies, suggest_repair, find_slskd_orphans) was parked
under lib/quality/ by the #477 monolith split even though it isn't
quality-decision logic. Pure move via `git mv` to a new top-level
lib/repair.py — function bodies are AST-identical; only module-level
docstrings/comments and callers' import paths changed. Removed the now
stale re-export block + __all__ entries from lib/quality/__init__.py
(single-operator repo, no compat shims) and repointed every caller
(lib/slskd_transfers.py, scripts/repair.py, tests/test_repair.py,
tests/test_repair_cli.py) at lib.repair directly.

Closes #512
